// frontend/src/api/generated/schemas/indexResearchResponse.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.
import type { MarketCoherenceSeries } from "./marketCoherenceSeries";
import type { IndexResearchParams } from "./indexResearchParams";
import type { SCIData } from "./sCIData";
import type { CoherenceSummary } from "./coherenceSummary";

export interface IndexResearchResponse {
  market: MarketCoherenceSeries;
  params: IndexResearchParams;
  sci: SCIData;
  summary: CoherenceSummary;
}